Move FromDecimal method to a separate file: Fraction.ConvertFromDecimal.cs (was a suggestion from @lipchev in a previous PR - which simplifies code reviews).
